Make it possible to update the pum after every typed character
If you want to add some completions to an already open pum you can set a
new variable to 1.

A use case would be if you want to use localcomplete after 2 keyword
chars, but don't want to add dictionary completions at that point
because that would add hundreds of words.  You can add the dictionary
completions at a later point, but if at that point the pum is still
open, the results will only be added when the new variable is 1.

Because it can be expensive (e.g.: RopeOmni) to check after every
character (and it is often useless) it can be set globally and
overridden per buffer:

    let g:acp_refeed_after_every_char = 1
    let b:acp_refeed_after_every_char = 0

Or vice versa.